+++ src/H5FDros3.c
@@ -34,9 +34,6 @@
 /* Define to turn on stats collection and reporting */
 /* #define ROS3_STATS */
 
-/* Max size of the cache, in bytes */
-#define ROS3_MAX_CACHE_SIZE 16777216
-
 /* The driver identification number, initialized at runtime */
 hid_t H5FD_ROS3_id_g = H5I_INVALID_HID;
 
@@ -51,8 +48,46 @@ static bool H5FD_ros3_init_s = false;
 /* Endpoint URL property name */
 #define ROS3_ENDPOINT_PROP_NAME "ros3_endpoint_prop"
 
+/* I/O block caching parameters property name */
+#define ROS3_BLOCK_CACHING_PARAMS_PROP_NAME "ros3_block_caching_params"
+
 /* Default page buffer size */
 #define ROS3_DEF_PAGE_BUF_SIZE ((size_t)64 * (size_t)1024 * (size_t)1024)
+
+/* Insert entry at head of LRU linked list */
+#define ROS3_BLOCK_CACHE_LRU_INSERT(file_ptr, block)                                                         \
+    do {                                                                                                     \
+        if (!(file_ptr)->block_cache.LRU_head) {                                                             \
+            (file_ptr)->block_cache.LRU_head = (block);                                                      \
+            (file_ptr)->block_cache.LRU_tail = (block);                                                      \
+        }                                                                                                    \
+        else {                                                                                               \
+            (file_ptr)->block_cache.LRU_head->prev = (block);                                                \
+            (block)->next                          = (file_ptr)->block_cache.LRU_head;                       \
+            (file_ptr)->block_cache.LRU_head       = (block);                                                \
+        }                                                                                                    \
+    } while (0)
+
+#define ROS3_BLOCK_CACHE_LRU_REMOVE(file_ptr, block)                                                         \
+    do {                                                                                                     \
+        if ((file_ptr)->block_cache.LRU_head == (block)) {                                                   \
+            (file_ptr)->block_cache.LRU_head = (block)->next;                                                \
+            if ((file_ptr)->block_cache.LRU_head)                                                            \
+                (file_ptr)->block_cache.LRU_head->prev = NULL;                                               \
+        }                                                                                                    \
+        else                                                                                                 \
+            (block)->prev->next = (block)->next;                                                             \
+                                                                                                             \
+        if ((file_ptr)->block_cache.LRU_tail == (block)) {                                                   \
+            (file_ptr)->block_cache.LRU_tail = (block)->prev;                                                \
+            if ((file_ptr)->block_cache.LRU_tail)                                                            \
+                (file_ptr)->block_cache.LRU_tail->next = NULL;                                               \
+        }                                                                                                    \
+        else                                                                                                 \
+            (block)->next->prev = (block)->prev;                                                             \
+                                                                                                             \
+        (block)->next = (block)->prev = NULL;                                                                \
+    } while (0)
 
 #ifdef ROS3_STATS
 
@@ -78,6 +113,34 @@ typedef struct H5FD_ros3_stats_bin {
 } H5FD_ros3_stats_bin_t;
 
 #endif /* ROS3_STATS */
+
+typedef struct H5FD_ros_block_hash_t {
+    UT_hash_handle hh; /* Hash table handle */
+
+    haddr_t addr;       /* Block size-aligned file address for block */
+    size_t  block_size; /* Size of block data buffer (buf) in bytes */
+
+    /* Fields for LRU eviction linked list */
+    struct H5FD_ros_block_hash_t *next;
+    struct H5FD_ros_block_hash_t *prev;
+
+    uint8_t buf[]; /* block data buffer; flexible array member */
+} H5FD_ros_block_hash_t;
+
+/* Structure for partitioning an I/O request into smaller requests
+ * along block size boundaries
+ */
+typedef struct H5FD_ros3_block_io_req_t {
+    haddr_t addr;
+    size_t  io_size;
+} H5FD_ros3_block_io_req_t;
+
+/* Parameters for I/O block caching */
+typedef struct H5FD_ros3_block_caching_params_t {
+    size_t block_size;
+    size_t block_cache_size;
+    bool   lock_superblock;
+} H5FD_ros3_block_caching_params_t;
 
 /***************************************************************************
  * Stores all information needed to maintain access to a single HDF5 file
@@ -105,14 +168,27 @@ typedef struct H5FD_ros3_stats_bin {
  *     Responsible for communicating with remote host and presenting file
  *     contents as indistinguishable from a file on the local filesystem.
  *
- * cache
- * cache_size (in bytes)
+ * block_cache
  *
- *     A simple cache of the first N bytes of the file. Especially useful
- *     at file open, when we perform several reads that would otherwise
- *     be uncached.
+ *     Holds fields for implementing a simple I/O block cache used for
+ *     optimizing I/O.
  *
- * *** present only if ROS3_SATS is set to enable stats collection ***
+ *     hash_table - Pointer to the head node of a uthash hash table that
+ *     is used for looking up cached I/O blocks by block size-aligned
+ *     addresses.
+ *
+ *     block_size - The size of each I/O block that is cached. Fixed upon
+ *     file open.
+ *
+ *     max_num_blocks - The maximum number of I/O blocks that can be cached
+ *     before blocks need to be evicted to make space for others.
+ *
+ *     lock_superblock - Boolean indicating whether the block which (usually)
+ *     contains the superblock and nearby metadata should be locked in the
+ *     cache to prevent its eviction. Note that a large userblock size
+ *     could prevent this approach from being effective.
+ *
+ * *** present only if ROS3_STATS is set to enable stats collection ***
  *
  * `meta` (H5FD_ros3_stats_bin_t[])
  * `raw` (H5FD_ros3_stats_bin_t[])
@@ -134,8 +210,18 @@ typedef struct H5FD_ros3_t {
     haddr_t          eoa;
     H5FD_ros3_fapl_t fa;
     s3r_t           *s3r_handle;
-    uint8_t         *cache;
-    size_t           cache_size;
+
+    struct {
+        H5FD_ros_block_hash_t *hash_table;
+        size_t                 block_size;
+        size_t                 block_cache_size;
+        size_t                 max_num_blocks;
+        bool                   lock_superblock;
+        H5FD_ros_block_hash_t *LRU_head;
+        H5FD_ros_block_hash_t *LRU_tail;
+        bool                   disabled;
+    } block_cache;
+
 #ifdef ROS3_STATS
     H5FD_ros3_stats_bin_t meta[ROS3_STATS_BIN_COUNT + 1];
     H5FD_ros3_stats_bin_t raw[ROS3_STATS_BIN_COUNT + 1];
@@ -172,6 +258,11 @@ static herr_t H5FD__ros3_str_endpoint_copy(const char *name, size_t size, void *
 static int    H5FD__ros3_str_endpoint_cmp(const void *_value1, const void *_value2, size_t size);
 static herr_t H5FD__ros3_str_endpoint_close(const char *name, size_t size, void *_value);
 static herr_t H5FD__ros3_str_endpoint_delete(hid_t prop_id, const char *name, size_t size, void *_value);
+
+static herr_t H5FD__ros3_init_block_cache(H5FD_ros3_t *file);
+static herr_t H5FD__ros3_determine_io_reqs(H5FD_ros3_t *file, haddr_t addr, size_t io_size,
+                                           H5FD_ros3_block_io_req_t **io_reqs_out, size_t *num_io_reqs_out);
+static herr_t H5FD__ros3_block_cache_make_space(H5FD_ros3_t *file);
 
 #ifdef ROS3_STATS
 static herr_t H5FD__ros3_reset_stats(H5FD_ros3_t *file);
@@ -931,6 +1022,144 @@ done:
 } /* end H5Pget_fapl_ros3_endpoint() */
 
 /*-------------------------------------------------------------------------
+ * Function:    H5FD__ros3_block_caching_params_cmp
+ *
+ * Purpose:     Compares two H5FD_ros3_block_caching_params_t structures
+ *
+ * Return:      -1/0/1
+ *-------------------------------------------------------------------------
+ */
+static int
+H5FD__ros3_block_caching_params_cmp(const void *value1, const void *value2, size_t H5_ATTR_UNUSED size)
+{
+    const H5FD_ros3_block_caching_params_t *params1 = (const H5FD_ros3_block_caching_params_t *)value1;
+    const H5FD_ros3_block_caching_params_t *params2 = (const H5FD_ros3_block_caching_params_t *)value2;
+
+    if (params1->block_size != params2->block_size)
+        return (params1->block_size > params2->block_size) - (params1->block_size < params2->block_size);
+    if (params1->block_cache_size != params2->block_cache_size)
+        return (params1->block_cache_size > params2->block_cache_size) -
+               (params1->block_cache_size < params2->block_cache_size);
+    if (params1->lock_superblock != params2->lock_superblock)
+        return (params1->lock_superblock > params2->lock_superblock) -
+               (params1->lock_superblock < params2->lock_superblock);
+
+    return 0;
+} /* end H5FD__ros3_block_caching_params_cmp() */
+
+/*-------------------------------------------------------------------------
+ * Function:    H5Pset_fapl_ros3_block_caching
+ *
+ * Purpose:     Sets I/O block caching parameters for the ros3 VFD.
+ *
+ * Return:      SUCCEED/FAIL
+ *-------------------------------------------------------------------------
+ */
+herr_t
+H5Pset_fapl_ros3_block_caching(hid_t fapl_id, size_t block_size, size_t block_cache_size,
+                               bool lock_superblock)
+{
+    H5FD_ros3_block_caching_params_t block_caching_params = {0};
+    H5P_genplist_t                  *plist                = NULL;
+    htri_t                           block_caching_params_exist;
+    herr_t                           ret_value = SUCCEED;
+
+    FUNC_ENTER_API(FAIL)
+
+    if (fapl_id == H5P_DEFAULT)
+        HGOTO_ERROR(H5E_PLIST, H5E_BADVALUE, FAIL, "can't set values in default property list");
+    if (NULL == (plist = H5P_object_verify(fapl_id, H5P_FILE_ACCESS, false)))
+        HGOTO_ERROR(H5E_PLIST, H5E_BADTYPE, FAIL, "not a file access property list");
+    if (H5FD_ROS3 != H5P_peek_driver(plist))
+        HGOTO_ERROR(H5E_PLIST, H5E_BADVALUE, FAIL, "ROS3 driver is not set on FAPL");
+
+    if ((block_caching_params_exist = H5P_exist_plist(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME)) < 0)
+        HGOTO_ERROR(H5E_PLIST, H5E_CANTGET, FAIL,
+                    "failed to check if I/O block caching parameters property exists in plist");
+
+    block_caching_params.block_size       = block_size;
+    block_caching_params.block_cache_size = block_cache_size;
+    block_caching_params.lock_superblock  = lock_superblock;
+
+    /* If block size is larger than block cache size, round block size down
+     * to block cache size so that block cache size is an upper limit but
+     * can still hold at least 1 block.
+     */
+    if (block_caching_params.block_size != 0 && block_caching_params.block_cache_size != 0) {
+        if (block_caching_params.block_size > block_caching_params.block_cache_size)
+            block_caching_params.block_size = block_caching_params.block_cache_size;
+    }
+
+    if (block_caching_params_exist) {
+        if (H5P_set(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME, &block_caching_params) < 0)
+            HGOTO_ERROR(H5E_PLIST, H5E_CANTSET, FAIL, "unable to set I/O block caching parameters");
+    }
+    else {
+        if (H5P_insert(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME, sizeof(H5FD_ros3_block_caching_params_t),
+                       &block_caching_params, NULL, NULL, NULL, NULL, NULL, NULL,
+                       H5FD__ros3_block_caching_params_cmp, NULL) < 0)
+            HGOTO_ERROR(H5E_PLIST, H5E_CANTREGISTER, FAIL,
+                        "unable to register I/O block caching parameters property in plist");
+    }
+
+done:
+    FUNC_LEAVE_API(ret_value)
+} /* end H5Pset_fapl_ros3_block_caching() */
+
+/*-------------------------------------------------------------------------
+ * Function:    H5Pget_fapl_ros3_block_caching
+ *
+ * Purpose:     Retrieves any I/O block caching parameters set for the ros3
+ *              VFD.
+ *
+ * Return:      SUCCEED/FAIL
+ *-------------------------------------------------------------------------
+ */
+herr_t
+H5Pget_fapl_ros3_block_caching(hid_t fapl_id, size_t *block_size, size_t *block_cache_size,
+                               bool *lock_superblock)
+{
+    H5P_genplist_t *plist = NULL;
+    htri_t          block_caching_params_exist;
+    herr_t          ret_value = SUCCEED;
+
+    FUNC_ENTER_API(FAIL)
+
+    if (NULL == (plist = H5P_object_verify(fapl_id, H5P_FILE_ACCESS, true)))
+        HGOTO_ERROR(H5E_PLIST, H5E_BADTYPE, FAIL, "not a file access property list");
+    if (H5FD_ROS3 != H5P_peek_driver(plist))
+        HGOTO_ERROR(H5E_PLIST, H5E_BADVALUE, FAIL, "ROS3 driver is not set on FAPL");
+
+    if ((block_caching_params_exist = H5P_exist_plist(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME)) < 0)
+        HGOTO_ERROR(H5E_PLIST, H5E_CANTGET, FAIL,
+                    "failed to check if I/O block caching parameters property exists in plist");
+    if (block_caching_params_exist) {
+        H5FD_ros3_block_caching_params_t block_caching_params = {0};
+
+        if (H5P_get(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME, &block_caching_params) < 0)
+            HGOTO_ERROR(H5E_PLIST, H5E_CANTGET, FAIL, "unable to get I/O block caching parameters");
+
+        if (block_size)
+            *block_size = block_caching_params.block_size;
+        if (block_cache_size)
+            *block_cache_size = block_caching_params.block_cache_size;
+        if (lock_superblock)
+            *lock_superblock = block_caching_params.lock_superblock;
+    }
+    else {
+        if (block_size)
+            *block_size = HDF5_ROS3_VFD_DEFAULT_BLOCK_SIZE;
+        if (block_cache_size)
+            *block_cache_size = ROS3_DEF_PAGE_BUF_SIZE;
+        if (lock_superblock)
+            *lock_superblock = true;
+    }
+
+done:
+    FUNC_LEAVE_API(ret_value)
+} /* end H5Pget_fapl_ros3_block_caching() */
+
+/*-------------------------------------------------------------------------
  * Function:    H5FD__ros3_open
  *
  * Purpose:     Create and/or open a file as an HDF5 file
@@ -952,14 +1181,15 @@ done:
 static H5FD_t *
 H5FD__ros3_open(const char *url, unsigned flags, hid_t fapl_id, haddr_t maxaddr)
 {
-    H5FD_ros3_t            *file          = NULL;
-    s3r_t                  *handle        = NULL;
-    const H5FD_ros3_fapl_t *fa            = NULL;
-    H5P_genplist_t         *plist         = NULL;
-    char                   *fapl_token    = NULL;
-    char                   *fapl_endpoint = NULL;
-    H5FD_t                 *ret_value     = NULL;
-    htri_t                  endpt_exists  = false;
+    H5FD_ros3_t            *file                       = NULL;
+    s3r_t                  *handle                     = NULL;
+    const H5FD_ros3_fapl_t *fa                         = NULL;
+    H5P_genplist_t         *plist                      = NULL;
+    char                   *fapl_token                 = NULL;
+    char                   *fapl_endpoint              = NULL;
+    H5FD_t                 *ret_value                  = NULL;
+    htri_t                  endpt_exists               = false;
+    htri_t                  block_caching_params_exist = false;
 
     FUNC_ENTER_PACKAGE
 
@@ -1022,22 +1252,42 @@ H5FD__ros3_open(const char *url, unsigned flags, hid_t fapl_id, haddr_t maxaddr)
     file->s3r_handle = handle;
     H5MM_memcpy(&(file->fa), fa, sizeof(H5FD_ros3_fapl_t));
 
+    if ((block_caching_params_exist = H5P_exist_plist(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME)) < 0)
+        HGOTO_ERROR(H5E_VFL, H5E_CANTGET, NULL,
+                    "failed to check if I/O block caching parameters property exists in plist");
+    if (block_caching_params_exist) {
+        H5FD_ros3_block_caching_params_t block_caching_params = {0};
+
+        if (H5P_get(plist, ROS3_BLOCK_CACHING_PARAMS_PROP_NAME, &block_caching_params) < 0)
+            HGOTO_ERROR(H5E_VFL, H5E_CANTGET, NULL, "unable to get I/O block caching parameters");
+
+        file->block_cache.block_size       = block_caching_params.block_size;
+        file->block_cache.block_cache_size = block_caching_params.block_cache_size;
+        file->block_cache.lock_superblock  = block_caching_params.lock_superblock;
+    }
+    else {
+        file->block_cache.block_size       = HDF5_ROS3_VFD_DEFAULT_BLOCK_SIZE;
+        file->block_cache.block_cache_size = HDF5_ROS3_VFD_DEFAULT_BLOCK_CACHE_SIZE;
+        file->block_cache.lock_superblock  = true;
+    }
+
+    if (file->block_cache.block_size == 0 || file->block_cache.block_cache_size == 0)
+        file->block_cache.disabled = true;
+
+    /* Determine the maximum number of blocks to keep around in the block cache */
+    if (!file->block_cache.disabled) {
+        /* final sanity check; the block size should never exceed the block cache size */
+        if (file->block_cache.block_size > file->block_cache.block_cache_size)
+            file->block_cache.block_size = file->block_cache.block_cache_size;
+
+        file->block_cache.max_num_blocks = file->block_cache.block_cache_size / file->block_cache.block_size;
+        assert(file->block_cache.max_num_blocks >= 1);
+    }
+
 #ifdef ROS3_STATS
     if (H5FD__ros3_reset_stats(file) < 0)
         HGOTO_ERROR(H5E_VFL, H5E_UNINITIALIZED, NULL, "unable to reset file statistics");
 #endif
-
-    /* Cache the initial bytes of the file */
-    {
-        size_t filesize = H5FD__s3comms_s3r_get_filesize(file->s3r_handle);
-
-        file->cache_size = (filesize < ROS3_MAX_CACHE_SIZE) ? filesize : ROS3_MAX_CACHE_SIZE;
-
-        if (NULL == (file->cache = (uint8_t *)H5MM_calloc(file->cache_size)))
-            HGOTO_ERROR(H5E_VFL, H5E_NOSPACE, NULL, "unable to allocate cache memory");
-        if (H5FD__s3comms_s3r_read(file->s3r_handle, 0, file->cache_size, file->cache, file->cache_size) < 0)
-            HGOTO_ERROR(H5E_VFL, H5E_READERROR, NULL, "unable to execute read");
-    }
 
     ret_value = (H5FD_t *)file;
 
@@ -1047,8 +1297,7 @@ done:
             if (H5FD__s3comms_s3r_close(handle) < 0)
                 HDONE_ERROR(H5E_VFL, H5E_CANTCLOSEFILE, NULL, "unable to close s3 file handle");
         if (file != NULL) {
-            file->cache = H5MM_xfree(file->cache);
-            file        = H5FL_FREE(H5FD_ros3_t, file);
+            file = H5FL_FREE(H5FD_ros3_t, file);
         }
     }
 
@@ -1084,8 +1333,17 @@ H5FD__ros3_close(H5FD_t H5_ATTR_UNUSED *_file)
         HGOTO_ERROR(H5E_VFL, H5E_CANTCLOSEFILE, FAIL, "unable to close S3 request handle");
 
     /* Release the file info */
-    file->cache = H5MM_xfree(file->cache);
-    file        = H5FL_FREE(H5FD_ros3_t, file);
+    if (file->block_cache.hash_table) {
+        H5FD_ros_block_hash_t *p, *tmp;
+
+        HASH_ITER(hh, file->block_cache.hash_table, p, tmp)
+        {
+            HASH_DEL(file->block_cache.hash_table, p);
+            H5MM_free(p);
+        }
+    }
+
+    file = H5FL_FREE(H5FD_ros3_t, file);
 
 done:
     FUNC_LEAVE_NOAPI(ret_value)
@@ -1333,14 +1591,18 @@ static herr_t
 H5FD__ros3_read(H5FD_t *_file, H5FD_mem_t H5_ATTR_UNUSED type, hid_t H5_ATTR_UNUSED dxpl_id, haddr_t addr,
                 size_t size, void *buf)
 {
-    H5FD_ros3_t *file      = (H5FD_ros3_t *)_file;
-    size_t       filesize  = 0;
-    herr_t       ret_value = SUCCEED;
+    H5FD_ros3_block_io_req_t *io_reqs       = NULL;
+    H5FD_ros_block_hash_t    *new_block     = NULL;
+    H5FD_ros3_t              *file          = (H5FD_ros3_t *)_file;
+    uint8_t                  *buf_ptr       = (uint8_t *)buf;
+    size_t                    filesize      = 0;
+    size_t                    num_io_blocks = 0;
+    bool                      is_cached     = false;
+    herr_t                    ret_value     = SUCCEED;
 
     FUNC_ENTER_PACKAGE
 
     assert(file);
-    assert(file->cache);
     assert(file->s3r_handle);
     assert(buf);
 
@@ -1349,13 +1611,68 @@ H5FD__ros3_read(H5FD_t *_file, H5FD_mem_t H5_ATTR_UNUSED type, hid_t H5_ATTR_UNU
     if ((addr > filesize) || ((addr + size) > filesize))
         HGOTO_ERROR(H5E_ARGS, H5E_OVERFLOW, FAIL, "range exceeds file address");
 
-    /* Copy from the cache when accessing the first N bytes of the file.
-     * Saves network I/O operations when opening files.
+    /* If this is the first read for the file and block caching is not disabled,
+     * read an initial "block size" worth of bytes and add the block to the block
+     * cache. This initial block is mostly used to optimize locating the file's
+     * superblock, as well as general metadata reads following that process.
+     * Note that while reading the superblock of a file, it can't be determined
+     * at that time whether paged aggregation will be enabled or not, so the
+     * first "block size" bytes (rather than just an estimated amount of superblock
+     * bytes) are always cached for possible later use. Also note that it can't
+     * be determined at that time whether a userblock exists or how large it is,
+     * so a userblock size >= "block size" will effectively make this cache
+     * useless.
      */
-    if (addr + size <= file->cache_size) {
-        memcpy(buf, file->cache + addr, size);
+    if (!file->block_cache.disabled && !file->block_cache.hash_table) {
+        if (H5FD__ros3_init_block_cache(file) < 0)
+            HGOTO_ERROR(H5E_VFL, H5E_CANTINIT, FAIL, "unable to initialize I/O block cache");
+        assert(file->block_cache.hash_table);
     }
-    else {
+
+    /* If the I/O request falls within the already-cached superblock block and
+     * either paged aggregation is enabled or the superblock block is locked
+     * in the cache, just serve the request from the cache. When paged aggregation
+     * is enabled, this can save a possible request to S3. When paged aggregation
+     * is not enabled, this saves a tiny bit of overhead from below since it's
+     * known that there will be only one I/O request block.
+     *
+     * Since no modifications are made to the block cache when paged aggregation
+     * is enabled other than caching the superblock block, the superblock block
+     * should always be available for use in this case. When paged aggregation
+     * isn't enabled, this optimization is only used when the superblock block
+     * is locked in the cache and thus should be found by a hash table lookup.
+     */
+    /* clang-format off */
+    is_cached =
+        !file->block_cache.disabled &&                               /* Block caching is enabled */
+        (addr + size <= file->block_cache.block_size) &&             /* Addr + size falls within cached block */
+        (file->pub.paged_aggr || file->block_cache.lock_superblock); /* Paged aggr. or locked cached block */
+    /* clang-format on */
+
+    if (is_cached) {
+        H5FD_ros_block_hash_t *super_block      = NULL;
+        haddr_t                super_block_addr = 0;
+
+        HASH_FIND(hh, file->block_cache.hash_table, &super_block_addr, sizeof(haddr_t), super_block);
+        if (!super_block)
+            HGOTO_ERROR(H5E_VFL, H5E_CANTGET, FAIL, "unable to locate superblock block in block cache");
+
+        /* No need to promote block in LRU - if paged aggregation is enabled,
+         * block caching (and thus the LRU eviction list) is not enabled.
+         * Otherwise, the superblock block is locked in the cache and not part
+         * of the LRU eviction policy.
+         */
+
+        memcpy(buf, super_block->buf + addr, size);
+        HGOTO_DONE(SUCCEED);
+    }
+
+    /* If block caching is disabled or if paged aggregation is enabled, just
+     * issue reads to S3 directly. When paged aggregation is enabled, it's
+     * assumed that higher layers are already performing caching and so
+     * additional caching is not needed here.
+     */
+    if (file->block_cache.disabled || file->pub.paged_aggr) {
         /*
          * Note that the VFD interface doesn't specify the size of buf.
          * Assume that the caller knows what they're doing.
@@ -1367,11 +1684,338 @@ H5FD__ros3_read(H5FD_t *_file, H5FD_mem_t H5_ATTR_UNUSED type, hid_t H5_ATTR_UNU
         if (H5FD__ros3_log_read_stats(file, type, (uint64_t)size) < 0)
             HGOTO_ERROR(H5E_VFL, H5E_CANTSET, FAIL, "unable to log read stats");
 #endif
+
+        HGOTO_DONE(SUCCEED);
+    }
+
+    /* If paged aggregation is not enabled, serve this I/O request from the block
+     * cache as able and otherwise issue reads to S3.
+     */
+
+    /* Split I/O request among block boundaries as necessary */
+    if (H5FD__ros3_determine_io_reqs(file, addr, size, &io_reqs, &num_io_blocks) < 0)
+        HGOTO_ERROR(H5E_VFL, H5E_READERROR, FAIL, "unable to partition read request into I/O blocks");
+
+    for (size_t i = 0; i < num_io_blocks; i++) {
+        H5FD_ros_block_hash_t *io_block   = NULL;
+        haddr_t                block_addr = HADDR_UNDEF;
+        bool                   can_cache  = true;
+
+        /* If block caching is enabled, check if the block is in the cache */
+        if (!file->block_cache.disabled) {
+            /* The first I/O request may not be "block size"-aligned; all others will be. */
+            if (i == 0)
+                block_addr = (io_reqs[i].addr / file->block_cache.block_size) * file->block_cache.block_size;
+            else
+                block_addr = io_reqs[i].addr;
+
+            HASH_FIND(hh, file->block_cache.hash_table, &block_addr, sizeof(haddr_t), io_block);
+            if (io_block) {
+                /* Serve read from the block cache if the block was found */
+                memcpy(buf_ptr, io_block->buf + (io_reqs[i].addr - io_block->addr), io_reqs[i].io_size);
+                buf_ptr += io_reqs[i].io_size;
+
+                /* If the block that was found is not the superblock block OR if it is
+                 * the superblock block and the block isn't locked in the cache (can be
+                 * evicted), promote the block to the head of the LRU eviction list
+                 * (making it a less likely candidate for eviction) if it's not already
+                 * at the head of the list.
+                 */
+                if (io_block->addr != 0 || !file->block_cache.lock_superblock) {
+                    /* At this point, the block should already be in the LRU eviction list. */
+                    assert(file->block_cache.LRU_head);
+                    assert((file->block_cache.LRU_head == io_block) || io_block->prev);
+
+                    if (file->block_cache.LRU_head != io_block) {
+                        ROS3_BLOCK_CACHE_LRU_REMOVE(file, io_block);
+                        ROS3_BLOCK_CACHE_LRU_INSERT(file, io_block);
+                    }
+                }
+
+                continue;
+            }
+        }
+
+        /* If the I/O request wasn't served from the block cache, check if the
+         * request can be cached before issuing it. A lack of space in the
+         * cache isn't considered here, as old blocks will be evicted as necessary.
+         */
+        if (file->block_cache.disabled)
+            can_cache = false;
+        else {
+            /* If the cache can only hold one block and the superblock block is
+             * locked in the cache, we cannot cache any more blocks.
+             */
+            if (file->block_cache.lock_superblock && file->block_cache.max_num_blocks == 1) {
+                assert(HASH_COUNT(file->block_cache.hash_table) == 1); /* superblock block should be cached */
+                can_cache = false;
+            }
+        }
+
+        /* If the I/O request can be cached, allocate a new block for it and insert
+         * it into the block cache after the request is finished.
+         */
+        if (can_cache) {
+            size_t alloc_size = sizeof(*new_block) + file->block_cache.block_size;
+            size_t read_size  = 0;
+
+            /* Allocate a new block and read the entire block's bytes */
+            if (NULL == (new_block = H5MM_malloc(alloc_size)))
+                HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "unable to allocate I/O block");
+            new_block->addr       = block_addr;
+            new_block->block_size = file->block_cache.block_size;
+            new_block->next       = NULL;
+            new_block->prev       = NULL;
+            memset(&new_block->hh, 0, sizeof(UT_hash_handle));
+
+            read_size = MIN(new_block->block_size, filesize - new_block->addr);
+            if (H5FD__s3comms_s3r_read(file->s3r_handle, new_block->addr, read_size, new_block->buf,
+                                       new_block->block_size) < 0)
+                HGOTO_ERROR(H5E_VFL, H5E_READERROR, FAIL, "unable to execute read");
+
+#ifdef ROS3_STATS
+            if (H5FD__ros3_log_read_stats(file, type, (uint64_t)read_size) < 0)
+                HGOTO_ERROR(H5E_VFL, H5E_CANTSET, FAIL, "unable to log read stats");
+#endif
+
+            memcpy(buf_ptr, new_block->buf + (io_reqs[i].addr - new_block->addr), io_reqs[i].io_size);
+
+            /* If block cache is full, evict oldest block before inserting a new block */
+            if (H5FD__ros3_block_cache_make_space(file) < 0)
+                HGOTO_ERROR(H5E_VFL, H5E_CANTFREE, FAIL, "unable to make space in block cache");
+
+            /* Add block to block cache */
+            HASH_ADD(hh, file->block_cache.hash_table, addr, sizeof(haddr_t), new_block);
+            if (!new_block->hh.tbl)
+                HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "unable to add I/O block to hash table");
+
+            /* Add block to head of LRU eviction list */
+            ROS3_BLOCK_CACHE_LRU_INSERT(file, new_block);
+
+            new_block = NULL; /* Now owned by hash table */
+        }
+        else {
+            /* Unable to cache this I/O request in the block cache. Just read
+             * from S3 directly. Note that the VFD interface doesn't specify
+             * the size of buf. Assume that the caller knows what they're doing.
+             */
+            if (H5FD__s3comms_s3r_read(file->s3r_handle, io_reqs[i].addr, io_reqs[i].io_size, buf_ptr,
+                                       io_reqs[i].io_size) < 0)
+                HGOTO_ERROR(H5E_VFL, H5E_READERROR, FAIL, "unable to execute read");
+
+#ifdef ROS3_STATS
+            if (H5FD__ros3_log_read_stats(file, type, (uint64_t)io_reqs[i].io_size) < 0)
+                HGOTO_ERROR(H5E_VFL, H5E_CANTSET, FAIL, "unable to log read stats");
+#endif
+        }
+
+        buf_ptr += io_reqs[i].io_size;
     }
 
 done:
+    H5MM_free(new_block);
+    H5MM_free(io_reqs);
+
     FUNC_LEAVE_NOAPI(ret_value)
 } /* end H5FD__ros3_read() */
+
+/*-------------------------------------------------------------------------
+ * Function:    H5FD__ros3_init_block_cache
+ *
+ * Purpose:     Initializes an I/O block cache for optimizing I/O. Reads
+ *              the first "block size" bytes of the file (or the entire
+ *              file if it's smaller than "block size") and stores the
+ *              block in a hash table for quick lookups by
+ *              "block size"-aligned addresses.
+ *
+ *              If the 'lock_superblock' field in the file struct is false,
+ *              this initial block will be added to an LRU list of blocks
+ *              which can be evicted when attempting to add another block
+ *              to the block cache. Otherwise, this block will be omitted
+ *              from that list, preventing it from being evicted until file
+ *              close.
+ *
+ * Return:      SUCCEED/FAIL
+ *-------------------------------------------------------------------------
+ */
+static herr_t
+H5FD__ros3_init_block_cache(H5FD_ros3_t *file)
+{
+    H5FD_ros_block_hash_t *super_block = NULL;
+    size_t                 file_size   = 0;
+    size_t                 alloc_size  = 0;
+    size_t                 read_size   = 0;
+    herr_t                 ret_value   = SUCCEED;
+
+    FUNC_ENTER_PACKAGE
+
+    assert(file);
+    assert(!file->block_cache.disabled);
+    assert(file->block_cache.block_size > 0);
+
+    /* Already initialized */
+    if (file->block_cache.hash_table)
+        HGOTO_DONE(SUCCEED);
+
+    alloc_size = sizeof(*super_block) + file->block_cache.block_size;
+
+    if (NULL == (super_block = H5MM_calloc(alloc_size)))
+        HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "unable to allocate I/O block");
+    super_block->addr       = 0;
+    super_block->block_size = file->block_cache.block_size;
+
+    /* Read either the entire file or "block size" bytes, whichever is smaller.
+     * If the file size is smaller than the "block size", the remaining bytes
+     * in the block will be zeroes.
+     */
+    file_size = H5FD__s3comms_s3r_get_filesize(file->s3r_handle);
+    read_size = (file_size < file->block_cache.block_size) ? file_size : file->block_cache.block_size;
+    if (H5FD__s3comms_s3r_read(file->s3r_handle, 0, read_size, super_block->buf,
+                               file->block_cache.block_size) < 0)
+        HGOTO_ERROR(H5E_VFL, H5E_READERROR, FAIL, "unable to execute read");
+
+    /* Add block to block cache */
+    HASH_ADD(hh, file->block_cache.hash_table, addr, sizeof(haddr_t), super_block);
+    if (!file->block_cache.hash_table)
+        HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "unable to allocate I/O block hash table");
+    if (!super_block->hh.tbl)
+        HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "unable to add I/O block to hash table");
+
+    /* Add the superblock block to the list of blocks that can be
+     * evicted if it isn't setup to be locked in the cache.
+     */
+    if (!file->block_cache.lock_superblock)
+        ROS3_BLOCK_CACHE_LRU_INSERT(file, super_block);
+
+    super_block = NULL; /* Now owned by hash table */
+
+done:
+    if (ret_value < 0) {
+        H5MM_free(super_block);
+    }
+
+    FUNC_LEAVE_NOAPI(ret_value)
+} /* end H5FD__ros3_init_block_cache() */
+
+/*-------------------------------------------------------------------------
+ * Function:    H5FD__ros3_determine_io_reqs
+ *
+ * Purpose:     Given an offset and length for an I/O request, partitions
+ *              the I/O request among "block size"-aligned boundaries and
+ *              returns a new array of smaller I/O requests.
+ *
+ * Return:      SUCCEED/FAIL
+ *-------------------------------------------------------------------------
+ */
+static herr_t
+H5FD__ros3_determine_io_reqs(H5FD_ros3_t *file, haddr_t addr, size_t io_size,
+                             H5FD_ros3_block_io_req_t **io_reqs_out, size_t *num_io_reqs_out)
+{
+    H5FD_ros3_block_io_req_t *io_reqs          = NULL;
+    haddr_t                   cur_addr         = HADDR_UNDEF;
+    haddr_t                   first_block_addr = HADDR_UNDEF;
+    haddr_t                   last_block_addr  = HADDR_UNDEF;
+    size_t                    block_size       = 0;
+    size_t                    num_blocks       = 0;
+    size_t                    num_blocks_left  = 0;
+    herr_t                    ret_value        = SUCCEED;
+
+    FUNC_ENTER_PACKAGE
+
+    assert(file);
+    assert(!file->block_cache.disabled);
+    assert(file->block_cache.block_size > 0);
+    assert(io_reqs_out);
+    assert(num_io_reqs_out);
+
+    if (io_size == 0) {
+        *io_reqs_out     = NULL;
+        *num_io_reqs_out = 0;
+        HGOTO_DONE(SUCCEED);
+    }
+
+    block_size = file->block_cache.block_size;
+
+    first_block_addr = ((addr / block_size) * block_size);
+    last_block_addr  = ((addr + io_size - 1) / block_size) * block_size;
+    num_blocks       = (last_block_addr / block_size + 1) - (first_block_addr / block_size);
+
+    assert(num_blocks > 0);
+    if (NULL == (io_reqs = H5MM_malloc(num_blocks * sizeof(*io_reqs))))
+        HGOTO_ERROR(H5E_VFL, H5E_CANTALLOC, FAIL, "couldn't allocate array of I/O requests");
+
+    /* Setup I/O request to first block */
+    io_reqs[0].addr    = addr;
+    io_reqs[0].io_size = MIN(io_size, block_size - (size_t)(addr - first_block_addr));
+    assert(io_reqs[0].io_size <= io_size);
+    io_size -= io_reqs[0].io_size;
+
+    num_blocks_left = num_blocks - 1;
+    cur_addr        = addr + io_reqs[0].io_size;
+
+    /* Setup I/O requests for any blocks between first and last */
+    if (num_blocks_left > 1) {
+        for (size_t i = 1; i < num_blocks - 1; i++) {
+            io_reqs[i].addr    = cur_addr;
+            io_reqs[i].io_size = block_size;
+
+            cur_addr += block_size;
+            assert(io_reqs[i].io_size <= io_size);
+            io_size -= io_reqs[i].io_size;
+
+            num_blocks_left--;
+        }
+    }
+
+    /* Setup I/O request for last block, if applicable */
+    if (num_blocks_left) {
+        assert(num_blocks_left == 1);
+        io_reqs[num_blocks - 1].addr    = cur_addr;
+        io_reqs[num_blocks - 1].io_size = io_size;
+        io_size -= io_reqs[num_blocks - 1].io_size;
+    }
+
+    assert(io_size == 0);
+
+    *io_reqs_out     = io_reqs;
+    *num_io_reqs_out = num_blocks;
+
+done:
+    FUNC_LEAVE_NOAPI(ret_value)
+} /* end H5FD__ros3_determine_io_reqs() */
+
+/*-------------------------------------------------------------------------
+ * Function:    H5FD__ros3_block_cache_make_space
+ *
+ * Purpose:     If necessary, evicts the oldest block in the block cache
+ *              (by LRU policy) to make space for a new block.
+ *
+ * Return:      SUCCEED/FAIL
+ *-------------------------------------------------------------------------
+ */
+static herr_t
+H5FD__ros3_block_cache_make_space(H5FD_ros3_t *file)
+{
+    H5FD_ros_block_hash_t *oldest_block = NULL;
+    herr_t                 ret_value    = SUCCEED;
+
+    FUNC_ENTER_PACKAGE_NOERR
+
+    assert(file);
+
+    /* If there's space in the cache, no need to evict a block */
+    if (HASH_COUNT(file->block_cache.hash_table) < file->block_cache.max_num_blocks)
+        HGOTO_DONE(SUCCEED);
+
+    oldest_block = file->block_cache.LRU_tail;
+    ROS3_BLOCK_CACHE_LRU_REMOVE(file, oldest_block);
+    HASH_DEL(file->block_cache.hash_table, oldest_block);
+
+    H5MM_free(oldest_block);
+
+done:
+    FUNC_LEAVE_NOAPI(ret_value)
+} /* end H5FD__ros3_block_cache_make_space() */
 
 /*-------------------------------------------------------------------------
  * Function:    H5FD__ros3_write